Windows build fix
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@44551 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/include/Performance.h
@@ -536,7 +536,7 @@ namespace pm
         ULONG getScale() { return 1; }
     private:
         com::Utf8Str mShortName;
-        ULONG        mSpeed;
+        uint32_t     mSpeed;
         SubMetric   *mLinkSpeed;
     };
